+++ maze_transformer/dataset/collected_dataset.py
@@ -25,6 +25,7 @@ from maze_transformer.dataset.maze_dataset import (
     MazeDatasetConfig,
     _coord_to_str,
 )
+from maze_transformer.utils.utils import corner_first_ndindex
 from maze_transformer.generation.constants import SPECIAL_TOKENS, Coord, CoordTup
 from maze_transformer.generation.lattice_maze import LatticeMaze
 
@@ -62,8 +63,8 @@ class MazeDatasetCollectionConfig(GPTDatasetConfig):
     def node_token_map(self) -> dict[CoordTup, str]:
         """map from node to token"""
         return {
-            tuple(coord): _coord_to_str(coord)
-            for coord in list(np.ndindex(self.max_grid_shape))
+            tuple(coord): _coord_to_str(coord) 
+            for coord in corner_first_ndindex(self.max_grid_n)
         }
 
     @cached_property
