+++ tests/unit/maze_transformer/utils/test_utils.py
@@ -1,8 +1,10 @@
 from pathlib import Path
 
+import numpy as np
+
 import pytest
 
-from maze_transformer.utils.utils import get_checkpoint_paths_for_run, register_method
+from maze_transformer.utils.utils import get_checkpoint_paths_for_run, register_method, corner_first_ndindex
 
 
 @pytest.mark.usefixtures("temp_dir")
@@ -48,3 +50,10 @@ def test_register_method():
     evalsB = TestEvalsB.evals
     assert list(evalsA.keys()) == ["eval_function"]
     assert list(evalsB.keys()) == ["other_eval_function"]
+
+def test_corner_first_ndindex():
+    for n in range(1, 11):
+        a_n = corner_first_ndindex(n)
+        a_n_plus_1 = corner_first_ndindex(n+1)
+        assert np.all(a_n == a_n_plus_1[:n**2])
+
